+++ at_functional_test/test/scan_verb_test.dart
@@ -14,7 +14,8 @@ void main() {
   Socket? socketFirstAtsign;
 
   test('Scan verb after authentication', () async {
-    var firstAtsignServer = ConfigUtil.getYaml()!['first_atsign_server']['first_atsign_url'];
+    var firstAtsignServer =
+        ConfigUtil.getYaml()!['first_atsign_server']['first_atsign_url'];
     socketFirstAtsign =
         await secure_socket_connection(firstAtsignServer, firstAtsignPort);
     socket_listener(socketFirstAtsign!);
@@ -35,7 +36,8 @@ void main() {
   }, timeout: Timeout(Duration(seconds: 120)));
 
   test('scan verb before authentication', () async {
-    var firstAtsignServer = ConfigUtil.getYaml()!['first_atsign_server']['first_atsign_url'];
+    var firstAtsignServer =
+        ConfigUtil.getYaml()!['first_atsign_server']['first_atsign_url'];
     socketFirstAtsign =
         await secure_socket_connection(firstAtsignServer, firstAtsignPort);
     socket_listener(socketFirstAtsign!);
@@ -48,7 +50,8 @@ void main() {
   }, timeout: Timeout(Duration(seconds: 120)));
 
   test('Scan verb with only atsign and no value', () async {
-    var firstAtsignServer = ConfigUtil.getYaml()!['first_atsign_server']['first_atsign_url'];
+    var firstAtsignServer =
+        ConfigUtil.getYaml()!['first_atsign_server']['first_atsign_url'];
     socketFirstAtsign =
         await secure_socket_connection(firstAtsignServer, firstAtsignPort);
     socket_listener(socketFirstAtsign!);
@@ -62,7 +65,8 @@ void main() {
   }, timeout: Timeout(Duration(seconds: 120)));
 
   test('Scan verb with regex', () async {
-    var firstAtsignServer = ConfigUtil.getYaml()!['first_atsign_server']['first_atsign_url'];
+    var firstAtsignServer =
+        ConfigUtil.getYaml()!['first_atsign_server']['first_atsign_url'];
     socketFirstAtsign =
         await secure_socket_connection(firstAtsignServer, firstAtsignPort);
     socket_listener(socketFirstAtsign!);
@@ -84,15 +88,16 @@ void main() {
   }, timeout: Timeout(Duration(seconds: 120)));
 
   test('Scan verb - Displays key with special characters', () async {
-    var firstAtsignServer = ConfigUtil.getYaml()!['first_atsign_server']['first_atsign_url'];
+    var firstAtsignServer =
+        ConfigUtil.getYaml()!['first_atsign_server']['first_atsign_url'];
     socketFirstAtsign =
         await secure_socket_connection(firstAtsignServer, firstAtsignPort);
     socket_listener(socketFirstAtsign!);
     await prepare(socketFirstAtsign!, firstAtsign);
 
     ///UPDATE VERB
-    await socket_writer(
-        socketFirstAtsign!, 'update:public:verifying,commas$firstAtsign Working?');
+    await socket_writer(socketFirstAtsign!,
+        'update:public:verifying,commas$firstAtsign Working?');
     var response = await read();
     print('update verb response : $response');
     assert(
@@ -103,6 +108,53 @@ void main() {
     response = await read();
     print('scan verb response : $response');
     expect(response, contains('"public:verifying,commas$firstAtsign"'));
+  }, timeout: Timeout(Duration(seconds: 120)));
+
+  test('Scan verb does not return expired keys', () async {
+    var firstAtsignServer =
+        ConfigUtil.getYaml()!['first_atsign_server']['first_atsign_url'];
+    socketFirstAtsign =
+        await secure_socket_connection(firstAtsignServer, firstAtsignPort);
+    socket_listener(socketFirstAtsign!);
+    await prepare(socketFirstAtsign!, firstAtsign);
+
+    ///UPDATE VERB
+    await socket_writer(
+        socketFirstAtsign!, 'update:public:verifyingTTL$firstAtsign Working?');
+    var response = await read();
+    print('update verb response : $response');
+    //update ttl so that key is expired
+    await socket_writer(socketFirstAtsign!,
+        'update:ttl:6000:public:verifyingTTL$firstAtsign 1');
+    sleep(Duration(seconds: 2));
+    await socket_writer(socketFirstAtsign!, 'scan');
+    response = await read();
+    print('scan verb response : $response');
+    expect(
+        false, response.contains('"update:public:verifyingTTL$firstAtsign"'));
+  }, timeout: Timeout(Duration(seconds: 120)));
+
+  test('Scan verb does not return unborn keys', () async {
+    var firstAtsignServer =
+        ConfigUtil.getYaml()!['first_atsign_server']['first_atsign_url'];
+    socketFirstAtsign =
+        await secure_socket_connection(firstAtsignServer, firstAtsignPort);
+    socket_listener(socketFirstAtsign!);
+    await prepare(socketFirstAtsign!, firstAtsign);
+
+    ///UPDATE VERB
+    await socket_writer(
+        socketFirstAtsign!, 'update:public:verifyingTTB$firstAtsign Working?');
+    var response = await read();
+    print('update verb response : $response');
+    //update ttb so that key is not born yet
+    await socket_writer(socketFirstAtsign!,
+        'update:ttb:6000:public:verifyingTTB$firstAtsign 600000');
+    await socket_writer(socketFirstAtsign!, 'scan');
+    response = await read();
+    print('scan verb response : $response');
+    expect(
+        false, response.contains('"update:public:verifyingTTB$firstAtsign"'));
   }, timeout: Timeout(Duration(seconds: 120)));
 
   tearDown(() {