+++ at_secondary/at_secondary_server/lib/src/verb/handler/notify_verb_handler.dart
@@ -2,7 +2,6 @@ import 'dart:collection';
 
 import 'package:at_commons/at_commons.dart';
 import 'package:at_persistence_secondary_server/at_persistence_secondary_server.dart';
-import 'package:at_persistence_secondary_server/src/notification/at_notification.dart';
 import 'package:at_secondary/src/connection/inbound/inbound_connection_metadata.dart';
 import 'package:at_secondary/src/notification/notification_manager_impl.dart';
 import 'package:at_secondary/src/notification/stats_notification_service.dart';
@@ -27,7 +26,8 @@ class NotifyVerbHandler extends AbstractVerbHandler {
       command.startsWith(getName(VerbEnum.notify) + ':') &&
       !command.startsWith('${getName(VerbEnum.notify)}:list') &&
       !command.startsWith('${getName(VerbEnum.notify)}:status') &&
-      !command.startsWith('${getName(VerbEnum.notify)}:all');
+      !command.startsWith('${getName(VerbEnum.notify)}:all') &&
+      !command.startsWith('${getName(VerbEnum.notify)}:remove');
 
   @override
   Verb getVerb() {
@@ -52,11 +52,15 @@ class NotifyVerbHandler extends AbstractVerbHandler {
     var ttr_ms;
     var ttln_ms;
     var isCascade;
+    // The notification id received from the SDK.
+    var id = verbParams['id'];
     var forAtSign = verbParams[FOR_AT_SIGN];
     var atSign = verbParams[AT_SIGN];
     var atValue = verbParams[AT_VALUE];
     atSign = AtUtils.formatAtSign(atSign);
     var key = verbParams[AT_KEY];
+    String? sharedKeyEncrypted = verbParams[SHARED_KEY_ENCRYPTED];
+    String? pubKeyCS = verbParams[SHARED_WITH_PUBLIC_KEY_CHECK_SUM];
     var messageType = SecondaryUtil().getMessageType(verbParams[MESSAGE_TYPE]);
     var strategy = verbParams[STRATEGY];
     // If strategy is null, default it to strategy all.
@@ -117,7 +121,7 @@ class NotifyVerbHandler extends AbstractVerbHandler {
       if (currentAtSign == forAtSign) {
         var notificationId = await NotificationUtil.storeNotification(
             forAtSign, atSign, key, NotificationType.received, opType,
-            value: atValue, ttl_ms: ttln_ms);
+            value: atValue, ttl_ms: ttln_ms, id: id);
         response.data = notificationId;
         return;
       }
@@ -138,6 +142,12 @@ class NotifyVerbHandler extends AbstractVerbHandler {
       if (ttl_ms != null) {
         atMetadata.ttl = ttl_ms;
       }
+      if (sharedKeyEncrypted != null) {
+        atMetadata.sharedKeyEnc = sharedKeyEncrypted;
+      }
+      if (pubKeyCS != null) {
+        atMetadata.pubKeyCS = pubKeyCS;
+      }
       final notificationBuilder = AtNotificationBuilder()
         ..fromAtSign = atSign
         ..toAtSign = forAtSign
@@ -157,6 +167,9 @@ class NotifyVerbHandler extends AbstractVerbHandler {
         ..atMetaData = atMetadata
         ..type = NotificationType.sent
         ..ttl = ttln_ms;
+      if(id != null && id.isNotEmpty){
+        notificationBuilder.id = id;
+      }
       var notificationId = await NotificationManager.getInstance()
           .notify(notificationBuilder.build());
       response.data = notificationId;
@@ -166,7 +179,7 @@ class NotifyVerbHandler extends AbstractVerbHandler {
       logger.info('Storing the notification $key');
       await NotificationUtil.storeNotification(
           fromAtSign, forAtSign, key, NotificationType.received, opType,
-          ttl_ms: ttln_ms, value: atValue);
+          ttl_ms: ttln_ms, value: atValue, id: id);
       // Setting isEncrypted variable to true. By default, value of all the keys are encrypted.
       // except for the public keys. So, if key is public set isEncrypted to false.
       var isEncrypted = true;
@@ -197,7 +210,9 @@ class NotifyVerbHandler extends AbstractVerbHandler {
                 ttb: ttb_ms,
                 ttr: ttr_ms,
                 ccd: isCascade,
-                isEncrypted: isEncrypted)
+                isEncrypted: isEncrypted,
+                sharedKeyEncrypted: sharedKeyEncrypted,
+                publicKeyChecksum: pubKeyCS)
             .build();
         cachedKeyCommitId =
             await _storeCachedKeys(key, metadata, atValue: atValue);