+++ at_secondary/at_secondary_server/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.0.15
+- Info verb now supports 'info:brief' usage
+## 3.0.14
+- Notify verb handler changes for shared key and public key checksum in metadata
+- Inbound connection management improvements
+- Update persistence version for hive upgrade
 ## 3.0.13
 - Changes to add responses to queue from last in outbound message listener
 - Uptake at_lookup version change for increase timeout for outbound connection